migrating /void and /refund to /transactions

+++ affirmpaymentcore/src/com/affirm/payment/service/executor/request/builder/AffirmRefundPaymentRequestServiceBuilder.java
@@ -2,11 +2,9 @@ package com.affirm.payment.service.executor.request.builder;
 
 import com.affirm.payment.AffirmPaymentConstants;
 
-import java.math.BigDecimal;
-
 public class AffirmRefundPaymentRequestServiceBuilder extends AffirmPaymentRequestServiceBuilder<AffirmRefundPaymentRequestServiceBuilder> {
 
-   public AffirmRefundPaymentRequestServiceBuilder setRefundAmount(final BigDecimal amount)
+   public AffirmRefundPaymentRequestServiceBuilder setRefundAmount(final int amount)
    {
       affirmPaymentServiceRequest.addParam(AffirmPaymentConstants.REFUND.AMOUNT, amount);
       return this;
